Feature: anomalyDetection으로 폴더이름 변경
