+++ src/main/java/com/stepguide/backend/domain/anomalyDetection/service/AnomalyDetectionService.java
@@ -1,10 +1,9 @@
-package com.stepguide.backend.domain.anomalydetection.service;
+package com.stepguide.backend.domain.anomalyDetection.service;
 
-import com.stepguide.backend.domain.anomalydetection.dto.AnomalyScoreDTO;
+import com.stepguide.backend.domain.anomalyDetection.dto.AnomalyScoreDTO;
 
 import java.math.BigDecimal;
 import java.time.LocalDateTime;
-import java.time.LocalTime;
 
 public interface AnomalyDetectionService {
     AnomalyScoreDTO calculateAnomalyDetectionScore(Integer userId, String accountNumber, LocalDateTime transactionTime, BigDecimal transactionAmount);
