+++ src/main/resources/mapper/anomalyDetection/AnomalyDetecionMapper.xml
@@ -2,7 +2,7 @@
 <!DOCTYPE mapper
         PUBLIC "-//mybatis.org//DTD Mapper 3.0//EN"
         "http://mybatis.org/dtd/mybatis-3-mapper.dtd">
-<mapper namespace="com.stepguide.backend.domain.anomalydetection.mapper.AnomalyDetectionMapper">
+<mapper namespace="com.stepguide.backend.domain.anomalyDetection.mapper.AnomalyDetectionMapper">
 
     <select id="countPrevioiusTransfers" resultType="java.lang.Integer">
         SELECT COUNT(*)
