+++ src/test/java/com/stepguide/backend/domain/anomalyDetection/mapper/AnomalyDetectionMapperTest.java
@@ -1,12 +1,10 @@
-package com.stepguide.backend.domain.anomalydetection.mapper;
+package com.stepguide.backend.domain.anomalyDetection.mapper;
 
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
 
 import java.time.LocalDate;
-
-import static org.junit.jupiter.api.Assertions.*;
 
 @SpringBootTest
 class AnomalyDetectionMapperTest {
